o3tl: ensure that the initializer of enumarray contains enough elements
Currently this silently succeeds. Turns out oox already contains some
too-short initializers, let's guess the missing properties are all
invalid.

One downside of the templated parameter pack approach in the enumarray ctor, as
witnessed in vcl/win/window/salframe.cxx, is that argument types can no longer
be implicitly deduced and thus need to be spelled explicitly now in certain
cases.

There were also three uses of enumarry with V being unsigned short (aka
sal_uInt16) that started to cause narrowing conversion errors now and needed to
be adapted:  In editeng/source/uno/unonrule.cxx the obvious fix was to use the
proper type for V.  In sw/source/core/unocore/unosett.cxx with its odd mix of
saL_Int16 and USHRT_MAX, lets keep things that way for now (probably awaiting
later clean up) and use casts to avoid the implicit narrowing.  And in
sw/source/filter/ww8/wrtw8esh.cxx the ESCHER_Prop_* values, while presumably
conceptionally of type sal_uInt16, are plain #defines (thus of type int), so
rather than changing V to int it looked more consistent to explicitly cast the
ESCHER_Prop_* vlaues to sal_uInt16.

(And in tools/source/fsys/urlobj.cxx the poor loplugin:redundantfcast started to
unhelpfully kick in for (only) the first argument now.)

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135970
Tested-by: Jenkins

+++ sw/source/core/unocore/unosett.cxx
@@ -242,12 +242,12 @@ static SwPageDesc* lcl_GetPageDesc(SwDoc* pDoc, const uno::Any& aValue)
 // Numbering
 const o3tl::enumarray<SvxAdjust, unsigned short> aSvxToUnoAdjust
 {
-    text::HoriOrientation::LEFT,   //3
-    text::HoriOrientation::RIGHT,  //1
-    USHRT_MAX,
-    text::HoriOrientation::CENTER, //2
-    USHRT_MAX,
-    USHRT_MAX
+    static_cast<unsigned short>(text::HoriOrientation::LEFT),   //3
+    static_cast<unsigned short>(text::HoriOrientation::RIGHT),  //1
+    static_cast<unsigned short>(USHRT_MAX),
+    static_cast<unsigned short>(text::HoriOrientation::CENTER), //2
+    static_cast<unsigned short>(USHRT_MAX),
+    static_cast<unsigned short>(USHRT_MAX)
 };
 
 const unsigned short aUnoToSvxAdjust[] =
